PBSolverCautious : Upperbound is a number of digits now
git-svn-id: svn://svn.forge.objectweb.org/svnroot/sat4j/maven/trunk@531 ba638df5-4473-46d1-82f8-c3ae2a17a6e1

+++ org.sat4j.pb/src/main/java/org/sat4j/pb/SolverFactory.java
@@ -27,8 +27,6 @@
  *******************************************************************************/
 package org.sat4j.pb;
 
-import java.math.BigInteger;
-
 import org.sat4j.core.ASolverFactory;
 import org.sat4j.minisat.core.IOrder;
 import org.sat4j.minisat.core.IPhaseSelectionStrategy;
@@ -362,8 +360,7 @@ public class SolverFactory extends ASolverFactory<IPBSolver> {
 	 *         PB-constraints to clauses is made in order to simplify cutting
 	 *         planes (if coefficients are larger than bound).
 	 */
-	public static PBSolverCautious newPBCPMixedConstraintsCautious(
-			BigInteger bound) {
+	public static PBSolverCautious newPBCPMixedConstraintsCautious(int bound) {
 		MiniSATLearning<PBDataStructureFactory> learning = new MiniSATLearning<PBDataStructureFactory>();
 		PBSolverCautious solver = new PBSolverCautious(new FirstUIP(),
 				learning, new PBMaxClauseCardConstrDataStructure(),
